--- a/sphinxcontrib/jsontable/adaptive/adaptive_control_integrator.py
+++ b/sphinxcontrib/jsontable/adaptive/adaptive_control_integrator.py
@@ -19,11 +19,9 @@ CLAUDE.md Code Excellence Compliance:
 - KISS原則: シンプル・直感的API設計
 """
 
-from dataclasses import dataclass
-from typing import Any, Dict, List, Optional
-import time
 import threading
-from datetime import datetime
+from dataclasses import dataclass
+from typing import Any, Dict
 
 
 @dataclass
@@ -244,7 +242,9 @@ class AdaptiveControlFoundationResult:
 
     def __post_init__(self):
         if self.adaptive_control_foundation_quality is None:
-            self.adaptive_control_foundation_quality = AdaptiveControlFoundationQuality()
+            self.adaptive_control_foundation_quality = (
+                AdaptiveControlFoundationQuality()
+            )
         if self.overall_adaptive_control_effect is None:
             self.overall_adaptive_control_effect = OverallAdaptiveControlEffect()
 
@@ -322,12 +322,14 @@ class AdaptiveControlIntegrator:
             "analytics_enhanced": True,
         }
 
-    def integrate_adaptive_control_components(self, options: Dict[str, Any]) -> ComponentIntegrationResult:
+    def integrate_adaptive_control_components(
+        self, options: Dict[str, Any]
+    ) -> ComponentIntegrationResult:
         """適応制御コンポーネント統合実装"""
         try:
             # コンポーネント統合処理実装
             integration_success = self._execute_component_integration(options)
-            
+
             if integration_success:
                 return ComponentIntegrationResult(
                     component_integration_success=True,
@@ -336,7 +338,7 @@ class AdaptiveControlIntegrator:
                 )
             else:
                 return self._handle_component_integration_error()
-                
+
         except Exception:
             return self._handle_component_integration_error()
 
@@ -347,14 +349,14 @@ class AdaptiveControlIntegrator:
             **self._integration_config,
             **options,
         }
-        
+
         # 統合効果計算
         integration_effectiveness = 0.92
         if integration_config.get("maximize_coordination_effects"):
             integration_effectiveness += 0.02
         if integration_config.get("synergy_optimization"):
             integration_effectiveness += 0.02
-            
+
         return integration_effectiveness >= 0.92
 
     def _handle_component_integration_error(self) -> ComponentIntegrationResult:
@@ -365,12 +367,14 @@ class AdaptiveControlIntegrator:
             synergy_optimization_active=True,
         )
 
-    def optimize_holistic_system_performance(self, options: Dict[str, Any]) -> HolisticOptimizationResult:
+    def optimize_holistic_system_performance(
+        self, options: Dict[str, Any]
+    ) -> HolisticOptimizationResult:
         """ホリスティックシステム最適化実装"""
         try:
             # ホリスティック最適化処理実装
             optimization_success = self._execute_holistic_optimization(options)
-            
+
             if optimization_success:
                 return HolisticOptimizationResult(
                     holistic_optimization_success=True,
@@ -379,7 +383,7 @@ class AdaptiveControlIntegrator:
                 )
             else:
                 return self._handle_holistic_optimization_error()
-                
+
         except Exception:
             return self._handle_holistic_optimization_error()
 
@@ -390,14 +394,14 @@ class AdaptiveControlIntegrator:
             **self._holistic_config,
             **options,
         }
-        
+
         # 最適化効果計算
         optimization_effectiveness = 0.95
         if optimization_config.get("system_wide_optimization"):
             optimization_effectiveness += 0.02
         if optimization_config.get("enterprise_grade_quality"):
             optimization_effectiveness += 0.01
-            
+
         return optimization_effectiveness >= 0.95
 
     def _handle_holistic_optimization_error(self) -> HolisticOptimizationResult:
@@ -408,12 +412,14 @@ class AdaptiveControlIntegrator:
             continuous_optimization_active=True,
         )
 
-    def coordinate_intelligent_adaptive_control(self, options: Dict[str, Any]) -> IntelligentCoordinationResult:
+    def coordinate_intelligent_adaptive_control(
+        self, options: Dict[str, Any]
+    ) -> IntelligentCoordinationResult:
         """インテリジェント適応協調実装"""
         try:
             # インテリジェント協調処理実装
             coordination_success = self._execute_intelligent_coordination(options)
-            
+
             if coordination_success:
                 return IntelligentCoordinationResult(
                     intelligent_coordination_success=True,
@@ -422,7 +428,7 @@ class AdaptiveControlIntegrator:
                 )
             else:
                 return self._handle_intelligent_coordination_error()
-                
+
         except Exception:
             return self._handle_intelligent_coordination_error()
 
@@ -433,14 +439,14 @@ class AdaptiveControlIntegrator:
             **self._coordination_config,
             **options,
         }
-        
+
         # 協調効果計算
         coordination_effectiveness = 0.89
         if coordination_config.get("ml_enhanced_control"):
             coordination_effectiveness += 0.03
         if coordination_config.get("predictive_control_integration"):
             coordination_effectiveness += 0.02
-            
+
         return coordination_effectiveness >= 0.89
 
     def _handle_intelligent_coordination_error(self) -> IntelligentCoordinationResult:
@@ -451,12 +457,14 @@ class AdaptiveControlIntegrator:
             predictive_control_integrated=True,
         )
 
-    def scale_distributed_adaptive_control(self, options: Dict[str, Any]) -> DistributedScalabilityResult:
+    def scale_distributed_adaptive_control(
+        self, options: Dict[str, Any]
+    ) -> DistributedScalabilityResult:
         """分散適応スケーラビリティ実装"""
         try:
             # 分散スケーラビリティ処理実装
             scalability_success = self._execute_distributed_scalability(options)
-            
+
             if scalability_success:
                 return DistributedScalabilityResult(
                     distributed_scaling_success=True,
@@ -465,7 +473,7 @@ class AdaptiveControlIntegrator:
                 )
             else:
                 return self._handle_distributed_scalability_error()
-                
+
         except Exception:
             return self._handle_distributed_scalability_error()
 
@@ -476,14 +484,14 @@ class AdaptiveControlIntegrator:
             **self._scalability_config,
             **options,
         }
-        
+
         # スケーラビリティ効果計算
         scalability_effectiveness = 0.95
         if scalability_config.get("high_availability_mode"):
             scalability_effectiveness += 0.02
         if scalability_config.get("distributed_coordination"):
             scalability_effectiveness += 0.01
-            
+
         return scalability_effectiveness >= 0.95
 
     def _handle_distributed_scalability_error(self) -> DistributedScalabilityResult:
@@ -494,12 +502,14 @@ class AdaptiveControlIntegrator:
             distributed_coordination_active=True,
         )
 
-    def guarantee_enterprise_adaptive_quality(self, options: Dict[str, Any]) -> EnterpriseQualityResult:
+    def guarantee_enterprise_adaptive_quality(
+        self, options: Dict[str, Any]
+    ) -> EnterpriseQualityResult:
         """企業グレード適応品質保証実装"""
         try:
             # 企業品質保証処理実装
             quality_success = self._execute_enterprise_quality_assurance(options)
-            
+
             if quality_success:
                 return EnterpriseQualityResult(
                     enterprise_quality_guaranteed=True,
@@ -508,7 +518,7 @@ class AdaptiveControlIntegrator:
                 )
             else:
                 return self._handle_enterprise_quality_error()
-                
+
         except Exception:
             return self._handle_enterprise_quality_error()
 
@@ -519,14 +529,14 @@ class AdaptiveControlIntegrator:
             **self._quality_config,
             **options,
         }
-        
+
         # 品質効果計算
         quality_effectiveness = 0.97
         if quality_config.get("quality_assurance_enforcement"):
             quality_effectiveness += 0.01
         if quality_config.get("audit_compliance_active"):
             quality_effectiveness += 0.01
-            
+
         return quality_effectiveness >= 0.97
 
     def _handle_enterprise_quality_error(self) -> EnterpriseQualityResult:
@@ -537,12 +547,14 @@ class AdaptiveControlIntegrator:
             sla_monitoring_active=True,
         )
 
-    def monitor_adaptive_control_performance(self, options: Dict[str, Any]) -> PerformanceMonitoringResult:
+    def monitor_adaptive_control_performance(
+        self, options: Dict[str, Any]
+    ) -> PerformanceMonitoringResult:
         """適応制御パフォーマンス監視実装"""
         try:
             # パフォーマンス監視処理実装
             monitoring_success = self._execute_performance_monitoring(options)
-            
+
             if monitoring_success:
                 return PerformanceMonitoringResult(
                     performance_monitoring_success=True,
@@ -551,7 +563,7 @@ class AdaptiveControlIntegrator:
                 )
             else:
                 return self._handle_performance_monitoring_error()
-                
+
         except Exception:
             return self._handle_performance_monitoring_error()
 
@@ -562,14 +574,14 @@ class AdaptiveControlIntegrator:
             **self._monitoring_config,
             **options,
         }
-        
+
         # 監視効果計算
         monitoring_effectiveness = 0.98
         if monitoring_config.get("realtime_monitoring_active"):
             monitoring_effectiveness += 0.01
         if monitoring_config.get("analytics_enhanced"):
             monitoring_effectiveness += 0.01
-            
+
         return monitoring_effectiveness >= 0.98
 
     def _handle_performance_monitoring_error(self) -> PerformanceMonitoringResult:
@@ -580,12 +592,16 @@ class AdaptiveControlIntegrator:
             continuous_optimization_enabled=True,
         )
 
-    def verify_integration_performance(self, options: Dict[str, Any]) -> IntegrationPerformanceResult:
+    def verify_integration_performance(
+        self, options: Dict[str, Any]
+    ) -> IntegrationPerformanceResult:
         """統合パフォーマンス検証実装"""
         try:
             # パフォーマンス検証処理実装
-            verification_success = self._execute_integration_performance_verification(options)
-            
+            verification_success = self._execute_integration_performance_verification(
+                options
+            )
+
             if verification_success:
                 return IntegrationPerformanceResult(
                     performance_verification_success=True,
@@ -594,22 +610,24 @@ class AdaptiveControlIntegrator:
                 )
             else:
                 return self._handle_integration_performance_error()
-                
+
         except Exception:
             return self._handle_integration_performance_error()
 
-    def _execute_integration_performance_verification(self, options: Dict[str, Any]) -> bool:
+    def _execute_integration_performance_verification(
+        self, options: Dict[str, Any]
+    ) -> bool:
         """統合パフォーマンス検証実行"""
         # GREEN実装: パフォーマンス検証処理
         verification_config = options
-        
+
         # パフォーマンススコア計算
         performance_score = 0.95
         if verification_config.get("minimize_control_overhead"):
             performance_score += 0.02
         if verification_config.get("high_efficiency_integration"):
             performance_score += 0.02
-            
+
         return performance_score >= 0.95
 
     def _handle_integration_performance_error(self) -> IntegrationPerformanceResult:
@@ -620,12 +638,14 @@ class AdaptiveControlIntegrator:
             overhead_minimized=True,
         )
 
-    def establish_adaptive_control_foundation(self, options: Dict[str, Any]) -> AdaptiveControlFoundationResult:
+    def establish_adaptive_control_foundation(
+        self, options: Dict[str, Any]
+    ) -> AdaptiveControlFoundationResult:
         """適応制御基盤確立実装"""
         try:
             # 基盤確立処理実装
             foundation_success = self._execute_foundation_establishment(options)
-            
+
             if foundation_success:
                 return AdaptiveControlFoundationResult(
                     foundation_establishment_success=True,
@@ -634,7 +654,7 @@ class AdaptiveControlIntegrator:
                 )
             else:
                 return self._handle_foundation_establishment_error()
-                
+
         except Exception:
             return self._handle_foundation_establishment_error()
 
@@ -642,14 +662,14 @@ class AdaptiveControlIntegrator:
         """基盤確立実行"""
         # GREEN実装: 基盤確立処理
         foundation_config = options
-        
+
         # 基盤品質スコア計算
         foundation_quality = 0.96
         if foundation_config.get("validate_overall_quality"):
             foundation_quality += 0.02
         if foundation_config.get("ensure_enterprise_grade_control"):
             foundation_quality += 0.01
-            
+
         return foundation_quality >= 0.96
 
     def _handle_foundation_establishment_error(self) -> AdaptiveControlFoundationResult: