… PERFORMANCE_OPTIMIZATION_MASTER_PLAN_FINAL_REPORT.md
… README.md
… docs/README.md
… docs/examples/user_friendly_error_examples.md
… docs/excel_advanced_features.md
… docs/getting_started.md
… docs/performance_optimization.md
… docs/troubleshooting_guide.md
… issues_to_create/github_issues_creation_commands.md
… issues_to_create/issue_documentation_user_experience.md
… issues_to_create/issue_merged_cells_detection.md
… issues_to_create/issue_range_detection_algorithms.md
… plan/Performance_Optimization_Master_Plan.md
… plan/progress/Performance_Optimization_Progress_Template.md
… plan/progress/daily_logs/2025-07-06.md
… plan/progress/daily_logs/2025-07-07.md
… plan/progress/daily_logs/2025-07-08.md
… pyproject.toml
… record/journal/daily_logs/2025-01-06.md
… record/journal/daily_logs/2025-07-07.md
… sphinxcontrib/jsontable/adaptive/auto_scaling_integrator.py
… sphinxcontrib/jsontable/adaptive/distributed_scaling_coordinator.py
… sphinxcontrib/jsontable/adaptive/dynamic_limit_configurator.py
… sphinxcontrib/jsontable/adaptive/load_detection_engine.py
… sphinxcontrib/jsontable/adaptive/metrics_collection_analyzer.py
… sphinxcontrib/jsontable/adaptive/monitoring_data_persistence.py
… sphinxcontrib/jsontable/adaptive/monitoring_integration_verification.py
… sphinxcontrib/jsontable/adaptive/network_bandwidth_adapter.py
… sphinxcontrib/jsontable/adaptive/performance_dashboard.py
… sphinxcontrib/jsontable/adaptive/processing_capacity_controller.py